\name{case.testP}
\alias{case.testP}
\title{
Print function for intercept/trend structure testing
}
\description{
Print function for intercept/trend structure testing
}
\usage{
case.testP(GVAR.o, sigLevel = 0.05)
}
\arguments{
  \item{GVAR.o}{object of class \code{"GVAR"}}
  \item{sigLevel}{significance level}
}
\details{
  used in the function \code{GVAR} if structural testing option is set to \code{TRUE}
}
\value{
  case test statistics are printed to the screen
}
\references{
  none
}
\author{
  Martin Feldkircher
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
  \code{\link{GVAR}}
}

\keyword{ts}
